Task_3_2_pro/app/requirements.txt
requests
datetime
pandas
psycopg2-binary
pathlib